tests: Test the log validator with fake faults
Provide a couple of example kernel faults, to exercise the kernel log
validator functionality.

The tests inject fake kernel faults into the kernel log buffer which
should be parsed by the validator.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

---
This test exercises the log validator of the test suite rather than
the kms subsystem itself.

As such - it is unlikely to be appropriate for integration as is, but
could be adapted as part of a test suite for the test suite.